refactor(chatbi): AskUserDialog 移到 components 目录
